[FIX] Long Function: renderEmailCredentialForm
Refactored renderEmailCredentialForm by extracting CSS, JS, and HTML templates into a separate file.
